fix: split tracks endpoint
